+++ app/src/main/java/com/example/workclass/MainActivity.kt
@@ -41,7 +41,9 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.workclass.ui.screens.HomeScreen
 import com.example.workclass.ui.screens.MainMenuScreen
+import com.example.workclass.ui.screens.TestScreen
 import com.example.workclass.ui.theme.WorkClassTheme
 
 class MainActivity : ComponentActivity() {
@@ -87,113 +89,8 @@ class MainActivity : ComponentActivity() {
     private fun column(function: () -> Unit) {
 
     }
-    @Preview(showBackground = true)
-    @Composable
-    fun TextComposable(name:String = "Vacio"){
-        Text("Weeeeelcoooomeeee")
-        Text(name)
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun ModifierExampleOne(){
-        Column (
-            modifier = Modifier
-                .padding(24.dp,10.dp,5.dp,20.dp)
 
-        ){
-            Text("My name is")
-        }
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun ModifierExampleTwo(){
-        Column (
-            modifier = Modifier
-                .padding(24.dp)
-                .fillMaxWidth()
-                .clickable(onClick={})
-        ){
-            Text("My name is")
-        }
-    }
-    fun clickAction(){
-        println("column clicked")
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun ModifierExampleThree(){
-        Column(
-            modifier = Modifier
-                .fillMaxHeight()
-                .padding(16.dp)
-                .background(Color.Cyan)
-                .border(width = 2.dp, Color.Black)
-                .width(200.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.SpaceEvenly
-        ) {
-            TextComposable("1.")
-            TextComposable("2")
-            TextComposable("3")
-            TextComposable("4")
-        }
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun ModifierExampleFour(){
-        Box(
-            modifier = Modifier
-                .background(Color.Green)
-                .padding(10.dp)
-                .height(300.dp)
-                .width(300.dp)
-        ){
-            Text("1",Modifier.align(Alignment.TopStart))
-            Text("2",Modifier.align(Alignment.TopCenter))
-            Text("3",Modifier.align(Alignment.TopEnd))
-            Text("4",Modifier.align(Alignment.CenterStart))
-            Text("5",Modifier.align(Alignment.Center))
-            Text("6",Modifier.align(Alignment.CenterEnd))
-            Text("7",Modifier.align(Alignment.BottomStart))
-            Text("8",Modifier.align(Alignment.BottomCenter))
-            Text("9",Modifier.align(Alignment.BottomEnd))
 
-        }
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun CustomText(){
-        Column() {
-            Text(stringResource(R.string.sample_text),
-                color = colorResource(R.color.teal_700),
-                fontSize = 10.sp,
-                fontStyle = FontStyle.Italic,
-                fontWeight = FontWeight.Thin
-            )
-            val gradientColor = listOf(Color.Cyan, Color.Magenta, Color.Green)
-            Text(
-                stringResource(R.string.sample_text),
-                style = TextStyle(Brush.linearGradient(colors = gradientColor)))
-        }
-    }
-    @Preview(showBackground = true)
-    @Composable
-    fun picture(){
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.Black)
-                .height(300.dp)
-        ){
-            Image(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                painter = painterResource(R.drawable.image),
-                contentDescription = "Logo Outlast",
-                contentScale = ContentScale.FillHeight
-            )
-        }
-    }
 
     @Composable
     fun ComposeMultiScreenApp(){
@@ -202,12 +99,13 @@ class MainActivity : ComponentActivity() {
     }
     @Composable
     fun SetupNavGraph(navController: NavHostController){
-        NavHost(navController = navController, startDestination = ""){
-            composable(""){
-                MainMenuScreen(navController)
+        NavHost(navController = navController, startDestination = "main_menu"){
+            composable("main_menu"){ MainMenuScreen(navController)}
+            composable("home_screen"){ HomeScreen(navController)}
+            composable("test_screen"){ TestScreen(navController)}
+
             }
         }
-    }
 
 
 }
