+++ app/src/main/java/com/example/workclass/ui/screens/MainMenuScreens.kt
@@ -3,6 +3,7 @@ package com.example.workclass.ui.screens
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -18,6 +19,12 @@ fun MainMenuScreen(navController: NavHostController){
             .fillMaxSize()
     ){
         Text("main menu")
+        Button(
+            onClick = {navController.navigate("home_screen")}
+        ) {
+            Text("Go to Home Screen")
+
+        }
 
 }
